# scrabbleScore-openapi.json
{"openapi":"3.1.0","info":{"title":"data access services scrabble score tracking","version":"1.0"},"servers":[{"url":"http://localhost:30000","description":"Generated server url"}],"paths":{"/player":{"get":{"tags":["scrabble-data-controller"],"summary":"get a list of all the available players","operationId":"getAllPlayers","responses":{"200":{"description":"list of players","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Player"}}}}}}},"post":{"tags":["scrabble-data-controller"],"summary":"add a new player","operationId":"addPlayer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Player"}}},"required":true},"responses":{"201":{"description":"newly added player","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Player"}}}}}}},"/game":{"get":{"tags":["scrabble-data-controller"],"summary":"get list of games","operationId":"getAllGames","responses":{"200":{"description":"list of games","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Game"}}}}}}},"post":{"tags":["scrabble-data-controller"],"summary":"add a new game","operationId":"addGame","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Game"}}},"required":true},"responses":{"201":{"description":"newly added game","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Game"}}}}}}},"/gamePlayer":{"post":{"tags":["scrabble-data-controller"],"summary":"create a game player","operationId":"addGamePlayer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GamePlayer"}}},"required":true},"responses":{"201":{"description":"the new game player","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GamePlayer"}}}}}}},"/gamePlayerRound":{"post":{"tags":["scrabble-data-controller"],"summary":"create a game player round","operationId":"addGamePlayerRound","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GamePlayerRound"}}},"required":true},"responses":{"201":{"description":"the new game player found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GamePlayerRound"}}}}}}},"/game/{id}":{"get":{"tags":["scrabble-data-controller"],"summary":"get specific game","operationId":"getGame","parameters":[{"name":"id","in":"path","description":"game ID","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"the game","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Game"}}}}}},"patch":{"tags":["scrabble-data-controller"],"summary":"mark a game complete","operationId":"markGameComplete","parameters":[{"name":"id","in":"path","description":"game ID","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json-patch+json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/JsonPatchSchema"}}}},"required":true},"responses":{"200":{"description":"the updated game","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Game"}}}}}}},"/player/{id}":{"get":{"tags":["scrabble-data-controller"],"summary":"get a specific player","operationId":"getPlayer","parameters":[{"name":"id","in":"path","description":"player ID","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"the player","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Player"}}}}}}},"/game/{gameId}/gamePlayer":{"get":{"tags":["scrabble-data-controller"],"summary":"get list of players for a game","operationId":"getGamePlayersForGame","parameters":[{"name":"gameId","in":"path","description":"game ID","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"the list of game players","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GamePlayer"}}}}}}}},"/game/{gameId}/gamePlayer/{sequenceId}":{"get":{"tags":["scrabble-data-controller"],"summary":"get a specific player in a game according to their order number","operationId":"getGamePlayerForGame","parameters":[{"name":"gameId","in":"path","description":"game ID","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"sequenceId","in":"path","description":"player sequence number within the game","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"the game player","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GamePlayer"}}}}}}}},"/game/{gameId}/gamePlayer/{sequenceId}/gamePlayerRound":{"get":{"tags":["scrabble-data-controller"],"summary":"get a game player'rounds within a game","operationId":"getGamePlayerRoundsForGamePlayer","parameters":[{"name":"gameId","in":"path","description":"game ID","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"sequenceId","in":"path","description":"player sequence number within the game","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"the game player rounds","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GamePlayerRound"}}}}}}}}},"components":{"schemas":{"Player":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"highGameId":{"type":"integer","format":"int32"}}},"Game":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"gameStatus":{"type":"string","enum":["Playing","Complete"]},"startTimestamp":{"type":"string","format":"date-time"},"endTimestamp":{"type":"string","format":"date-time"},"winnerPlayerId":{"type":"integer","format":"int32"}}},"GamePlayer":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"gameId":{"type":"integer","format":"int32"},"playerId":{"type":"integer","format":"int32"},"score":{"type":"integer","format":"int32"},"sequenceNumber":{"type":"integer","format":"int32"}}},"GamePlayerRound":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"gamePlayerId":{"type":"integer","format":"int32"},"score":{"type":"integer","format":"int32"},"sevenLetter":{"type":"boolean"},"notes":{"type":"string"},"round":{"type":"integer","format":"int32"},"rollingScore":{"type":"integer","format":"int32"}}},"JsonPatchSchema":{"type":"object","properties":{"op":{"type":"string","enum":["replace","add","remove","copy","move","test"],"minLength":1},"path":{"type":"string","example":"/name","minLength":1},"value":{"type":"string","minLength":1}}}}}}